Add logo and working with orders_controller

+++ app/helpers/orders_helper.rb
@@ -2,6 +2,17 @@
 
 module OrdersHelper
 
+	def change_mystock(deal)
+		# Если у покупателя пока нет ни одной акции эмитента акций
+		if !current_user.mystocks.where(:host_id == deal.host_id).first
+			Mystock.create(:user_id => deal.user_id, :host_id => deal.host_id, :price => deal.price, :amount => deal.amount)
+		# Если у пользователя уже есть акции этого эмитента
+		else
+			stock = current_user.mystocks.where(:host_id == deal.host_id).first
+			stock.update_attributes(:amount => stock.amount + deal.amount, :price => deal.price)
+		end
+	end
+
 	def make_deal(second_user, host_id, status, amount, price)
 		@deal = current_user.deals.build(
 			:host_id => host_id,
@@ -12,14 +23,20 @@ module OrdersHelper
 			)
 
 		@deal.save
+
+		# Вносим изменения в имеющиеся у пользователя акции
+		change_mystock(@deal)
 	end
 
 	def compare_orders(new_order, status)
-		Order.where("host_id = ? AND status = ? AND price <= ?", new_order.host_id, status, new_order.price).each do |old_order|
+		for old_order in Order.where("host_id = ? AND status = ? AND price >= ?", new_order.host_id, status, new_order.price)
+  			unless new_order.frozen?
+
 			if old_order.amount > new_order.amount 
 				amount = old_order.amount - new_order.amount
 
-				make_deal(old_order.user.id, new_order.host_id, 1, amount, new_order.price)
+				# Создаём сделку
+				make_deal(old_order.user.id, new_order.host_id, status, new_order.amount, new_order.price)
 
 				old_order.update_attribute(:amount, amount)
 				new_order.update_attribute(:amount, 0)
@@ -28,7 +45,8 @@ module OrdersHelper
 
 			elsif old_order.amount == new_order.amount 
 
-				make_deal(old_order.user.id, new_order.host_id, 1, new_order.amount, new_order.price)
+				# Создаём сделку
+				make_deal(old_order.user.id, new_order.host_id, status, new_order.amount, new_order.price)
 
 				old_order.update_attribute(:amount, 0)
 				new_order.update_attribute(:amount, 0)
@@ -36,16 +54,18 @@ module OrdersHelper
 				old_order.destroy
 				new_order.destroy
 
-			else old_order.amount < new_order.amount
+			elsif old_order.amount < new_order.amount
 				amount = new_order.amount - old_order.amount
 
-				make_deal(old_order.user.id, new_order.host_id, 1, amount, new_order.price)
+				# Создаём сделку
+				make_deal(old_order.user.id, new_order.host_id, status, amount, new_order.price)
 
 				old_order.update_attribute(:amount, 0)
 				new_order.update_attribute(:amount, amount)
 
 				old_order.destroy
 			end							
+		end
 		end
 	end
